service: Move AirblocBackend to service package
